feat(art): scrolls replace coins, dinosaurs a quarter larger, rex on the icon
- collectible is a golden scroll with a red ribbon; hud icon and map text follow
- enemy sprites and hitboxes scaled by 1.25 through a fractional sprite scale

+++ Sources/Ninjasaurus/Rendering/PixelPainter.swift
@@ -195,7 +195,7 @@ struct PixelPainter {
         return out
     }
 
-    func sprite(scale: Int = 2) -> PixelSprite {
+    func sprite(scale: Double = 2) -> PixelSprite {
         PixelSprite(width: width, height: height, pixels: pixels, scale: scale)
     }
 }
